* develop(1105-sftp-support): adding SFTP support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): fix SSM parameter import/export

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add SFTP server

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add SFTP server

* develop(1105-sftp-support): add Cognito identity provider details

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add Cognito identity provider details

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add Cognito identity provider details

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* Merge remote-tracking branch 'refs/remotes/origin/main' into 1105-sft…

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): change logging

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* develop(1105-sftp-support): fix SQS message attributes

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): add FTP users protocols

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix queue/topic import

* develop(1105-sftp-support): fix awslocal

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* Merge remote-tracking branch 'origin/1105-sftp-support' into 1105-sft…

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

* develop(1105-sftp-support): new import request

* Merge branch 'main' into 1105-sftp-support

* develop(1105-sftp-support): new import request

+++ src/db/include/awsmock/repository/SQSDatabase.h
@@ -361,7 +361,7 @@ namespace AwsMock::Database {
          * @param maxRetries maximal number of retries
          * @param messageList message list
          */
-        void ReceiveMessages(const std::string &queueArn, int visibility, int maxResult, const std::string &dlQueueArn, int maxRetries, Entity::SQS::MessageList &messageList) const;
+        void ReceiveMessages(const std::string &queueArn, long visibility, long maxResult, const std::string &dlQueueArn, long maxRetries, Entity::SQS::MessageList &messageList) const;
 
         /**
          * @brief Reset expired resources
